- Added settings to choose between native compiler and legacy; - Added help file on compilers; - Optimized some project files;
