+++ src/NWScriptCompiler.h
@@ -10,7 +10,9 @@
 #include <string>
 #include <vector>
 
-#include "Nsc.h"
+#include "Native Compiler/exobase.h"		// New oficial compiler provided by Beamdog itself.
+#include "Native Compiler/scriptcomp.h"		// 
+#include "Nsc.h"							// Here we are using NscLib for older features like preprocessor and make dependency
 #include "Common.h"
 
 #include "Settings.h"
@@ -18,12 +20,53 @@
 
 namespace NWScriptPlugin
 {
+
+	// Copied from NscCompiler.cpp -> Resource Cache structures
+	struct ResourceCacheKey
+	{
+		NWN::ResRef32 ResRef;
+		NWN::ResType  ResType;
+
+		inline bool operator < (const ResourceCacheKey& other) const
+		{
+			return (ResType < other.ResType) ||
+				(memcmp(&ResRef, &other.ResRef, sizeof(ResRef)) < 0);
+		}
+
+		inline bool operator == (const ResourceCacheKey& other) const
+		{
+			return (ResType == other.ResType) &&
+				(memcmp(&ResRef, &other.ResRef, sizeof(ResRef)) == 0);
+		}
+	};
+
+	struct ResourceCacheEntry
+	{
+		bool                Allocated;
+		char*				Contents;
+		UINT32              Size;
+		std::string         Location;
+	};
+
+	typedef std::map<ResourceCacheKey, ResourceCacheEntry> ResourceCache;
+
+	struct NativeCompileResult
+	{
+		int32_t code;
+		char* str; // static buffer
+	};
+
+	// Function pointers to resolve new compiler Resource API requirements
+	static BOOL ResManUpdateResourceDirectory(const char* sAlias);
+	static int32_t ResManWriteToFile(const char* sFileName, RESTYPE nResType, const uint8_t* pData, size_t nSize, bool bBinary);
+	static const char* ResManLoadScriptSourceFile(const char* sFileName, RESTYPE nResType);
+	static const char* TlkResolve(STRREF strRef);
+
 	class NWScriptCompiler final
 	{
 	public:
 
-		NWScriptCompiler() : 
-			_resourceManager(nullptr), _settings(nullptr), _compiler(nullptr) {}
+		NWScriptCompiler();
 
 		bool isInitialized() {
 			return _resourceManager != nullptr;
@@ -38,18 +81,7 @@ namespace NWScriptPlugin
 		bool initialize();
 
 		// Reset compiler state
-		void reset() {
-			_resourceManager = nullptr;
-			_compiler = nullptr;
-			_includePaths.clear();
-			_fetchPreprocessorOnly = false;
-			_makeDependencyView = false;
-			_sourcePath = "";
-			_destDir = "";
-			setMode(0);
-			_processingEndCallback = nullptr;
-			clearLog();
-		}
+		void reset();
 
 		// Sets destination to a VALID and existing directory (or else get an error)
 		void setDestinationDirectory(fs::path dest) {
@@ -112,15 +144,15 @@ namespace NWScriptPlugin
 			_makeDependencyView = false;
 		}
 
-		int getMode() const {
+		inline int getMode() const {
 			return _compilerMode;
 		}
 
-		bool isViewDependencies() const {
+		inline bool isViewDependencies() const {
 			return _makeDependencyView;
 		}
 
-		bool isFetchPreprocessorOnly() const {
+		inline bool isFetchPreprocessorOnly() const {
 			return _fetchPreprocessorOnly;
 		}
 
@@ -128,17 +160,31 @@ namespace NWScriptPlugin
 			return _logger;
 		}
 
+		std::vector<std::string>& includePaths() {
+			return _includePaths;
+		}
+
 		// Returns if an output path is required for operation
-		bool isOutputDirRequired() {
+		inline bool isOutputDirRequired() {
 			return !(_fetchPreprocessorOnly || _makeDependencyView);
+		}
+
+		inline ResourceCache& getResourceCache() {
+			return _ResourceCache;
 		}
 
 
 		void processFile(bool fromMemory, char* fileContents);
 
 	private:
+
 		std::unique_ptr<ResourceManager> _resourceManager;
-		std::unique_ptr<NscCompiler> _compiler;
+		ResourceCache _ResourceCache;
+		std::unique_ptr<CScriptCompiler> _compilerNative;
+
+		// # TODO: Remove old compiler references
+		std::unique_ptr<NscCompiler> _compilerLegacy;
+
 		bool _fetchPreprocessorOnly = false;
 		bool _makeDependencyView = false;
 		int _compilerMode = 0;
@@ -163,7 +209,10 @@ namespace NWScriptPlugin
 		bool loadScriptResources();
 
 		// Compile a plain text script into binary format
-		bool compileScript(std::string& fileContents,
+		bool compileScriptLegacy(std::string& fileContents,
+			const NWN::ResType& fileResType, const NWN::ResRef32& fileResRef);
+
+		bool compileScriptNative(std::string& fileContents,
 			const NWN::ResType& fileResType, const NWN::ResRef32& fileResRef);
 
 		// Disassemble a binary file into a pcode assembly text format
