+++ src/NWScriptLogger.cpp
@@ -12,22 +12,22 @@
 #include "NWScriptLogger.h"
 
 #define PREPROCESSORPARSE R"((?:[\w\s\\.\-\(\):]+)(?:[\w\s\\.\-\(\):]+)\((?:\d+)\):\s(?:\w+):\s(?:NSC6022): Preprocessed: (.*))"
-#define COMPILERREGEX R"((?<fileName>[\w\s\\.\-\(\):]+)\.(?<fileExt>[\w\s\\.\-\(\):]+)\((?<lineNumber>\d+)\):\s(?<type>\w+):\s(?<code>NSC\d+):\s(?<message>.+))"
-#define COMPILERREGEXV2 R"((?<fileName>[\w\s\\.\-\(\):]+)\.(?<fileExt>[\w\s\\.\-\(\):]+)\((?<lineNumber>\d+)\):\s(?<type>\w+):\s(?<message>.+)\s(?<code>NSC\d+))"
+#define COMPILERREGEXNATIVE R"((?<fileName>[\w\s\\.\-\(\):]+)\.(?<fileExt>[\w\s\\.\-\(\):]+)\((?<lineNumber>\d+)\):\s(?<type>\w+):\s(?<message>.+)\s(?<code>NSC\d+))"
+#define COMPILERREGEXLEGACY R"((?<fileName>[\w\s\\.\-\(\):]+)\.(?<fileExt>[\w\s\\.\-\(\):]+)\((?<lineNumber>\d+)\):\s(?<type>\w+):\s(?<code>NSC\d+):\s(?<message>.+))"
 #define GENERALMESSAGE R"(\s*(?<type>WARNING|ERROR|INFO)\s*:\s*(?<message>.+))"
 #define INCLUDESPARSEREGEX R"( ShowIncludes: Handled resource ([^\/]+)\/([^\\\n]+))"
 
 typedef jpcre2::select<char> pcre2;
 
 static const pcre2::Regex preprocessorRegex(PREPROCESSORPARSE, 0, jpcre2::JIT_COMPILE);
-static const pcre2::Regex fileParsingMessageRegex(COMPILERREGEX, 0, jpcre2::JIT_COMPILE);
-static const pcre2::Regex fileParsingMessageRegexV2(COMPILERREGEXV2, 0, jpcre2::JIT_COMPILE);
+static const pcre2::Regex fileParsingMessageRegexNative(COMPILERREGEXNATIVE, 0, jpcre2::JIT_COMPILE);
+static const pcre2::Regex fileParsingMessageRegexLegacy(COMPILERREGEXLEGACY, 0, jpcre2::JIT_COMPILE);
 static const pcre2::Regex generalMessageRegex(GENERALMESSAGE, PCRE2_CASELESS, jpcre2::JIT_COMPILE);
 static const pcre2::Regex includesRegex(INCLUDESPARSEREGEX, 0, jpcre2::JIT_COMPILE);
 
 static pcre2::RegexMatch preprocessorParsing(&preprocessorRegex);
-static pcre2::RegexMatch fileParsingMessage(&fileParsingMessageRegex);
-static pcre2::RegexMatch fileParsingMessageV2(&fileParsingMessageRegexV2);
+static pcre2::RegexMatch fileParsingMessageNative(&fileParsingMessageRegexNative);
+static pcre2::RegexMatch fileParsingMessageLegacy(&fileParsingMessageRegexLegacy);
 static pcre2::RegexMatch generalMessage(&generalMessageRegex);
 static pcre2::RegexMatch includeFile(&includesRegex);
 
@@ -84,9 +84,11 @@ void NWScriptLogger::WriteTextV(const char* fmt, va_list ap)
 	}
 
 	// Check for compiler file parsing messages
-	fileParsingMessage.setNamedSubstringVector(&namedGroup);
-	fileParsingMessage.setSubject(buf);
-	matches = fileParsingMessage.match();	
+
+	// Parsing messages from new Beamdog's compiler
+	fileParsingMessageNative.setNamedSubstringVector(&namedGroup);
+	fileParsingMessageNative.setSubject(buf);
+	matches = fileParsingMessageNative.match();
 	if (matches)
 	{
 		log(namedGroup[0]["message"],
@@ -96,10 +98,10 @@ void NWScriptLogger::WriteTextV(const char* fmt, va_list ap)
 		return;
 	}
 
-	// Parsing messages from new Beamdog's compiler
-	fileParsingMessageV2.setNamedSubstringVector(&namedGroup);
-	fileParsingMessageV2.setSubject(buf);
-	matches = fileParsingMessageV2.match();
+	// Parsing messages from Legacy Compiler
+	fileParsingMessageLegacy.setNamedSubstringVector(&namedGroup);
+	fileParsingMessageLegacy.setSubject(buf);
+	matches = fileParsingMessageLegacy.match();	
 	if (matches)
 	{
 		log(namedGroup[0]["message"],
