+++ src/Plugin Controls/WarningDialog.cpp
