+++ src/Plugin Controls/WarningDialog.h
